refactor: default values for prod

+++ app/tasks/page.tsx
@@ -1,7 +1,9 @@
 "use client";
 import AppShell from "@/components/app-shell";
 import { useAuthGuard } from "@/components/auth-guard";
+import { getUser } from "@/lib/utils";
 import { User, Task } from "@prisma/client";
+import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 
 type TaskWithUser = Task & { assignee: User | null };
@@ -12,9 +14,9 @@ export default function TasksPage() {
   const [currentPage, setCurrentPage] = useState(1);
   const [total, setTotal] = useState(0);
   const [limit, setLimit] = useState(10);
-
+  const router = useRouter();
   useAuthGuard();
-
+  const user = getUser();
   const fetchTasks = async () => {
     setLoading(true);
     try {
@@ -29,15 +31,17 @@ export default function TasksPage() {
     }
   };
   const handleDelete = async (taskId: number) => {
-    console.log("taskId", taskId);
+    if (user?.role !== "Owner") {
+      alert("Only Owner and Admin can delete tasks!!");
+      return;
+    }
     const confirmed = confirm("Are you sure you want to delete this task?");
     if (!confirmed) return;
-
     try {
       const res = await fetch(`/api/tasks/${taskId}`, { method: "DELETE" });
       if (!res.ok) throw new Error("Failed to delete task");
 
-      setCurrentPage(1)
+      setCurrentPage(1);
       alert("Task deleted successfully");
     } catch (err) {
       console.error(err);
@@ -56,12 +60,26 @@ export default function TasksPage() {
       title="Tasks"
       subtitle="Create, update, delete, and assign tasks"
       actions={
-        <a
-          href="/tasks/new"
-          className="inline-flex items-center rounded-md bg-primary px-3 py-2 text-primary-foreground text-sm hover:opacity-90"
-        >
-          Create Task
-        </a>
+        <>
+          <a
+            href="/tasks/new"
+            className="inline-flex items-center rounded-md bg-primary px-3 py-2 text-primary-foreground text-sm hover:opacity-90"
+          >
+            Create Task
+          </a>
+
+          <button
+            onClick={() => {
+              localStorage.removeItem("token");
+              sessionStorage.removeItem("token");
+
+              router.push("/login");
+            }}
+            className="inline-flex items-center rounded-md bg-primary px-3 py-2 text-primary-foreground text-sm hover:opacity-90"
+          >
+            Logout
+          </button>
+        </>
       }
     >
       <div className="rounded-lg border border-border bg-card overflow-hidden">
@@ -72,7 +90,7 @@ export default function TasksPage() {
                 <th>Title</th>
                 <th>Assignee</th>
                 <th>Status</th>
-                <th>Time To Complete</th>
+                {/* <th>Time To Complete</th> */}
                 <th>Priority</th>
                 <th className="text-right">Actions</th>
               </tr>
@@ -89,10 +107,10 @@ export default function TasksPage() {
                   </td>
                   <td>
                     <span className="rounded-full bg-accent px-2 py-1 text-xs">
-                      {task.status}
+                      {task.status.replace("_", " ")}
                     </span>
                   </td>
-                  <td>5 min</td>
+                  {/* <td>5 min</td> */}
                   <td>
                     <span className="rounded-full bg-secondary px-2 py-1 text-xs">
                       {task.priority}
@@ -100,15 +118,30 @@ export default function TasksPage() {
                   </td>
                   <td className="text-right">
                     <div className="inline-flex gap-2">
-                      <a
-                        href={`/tasks/${task.id}/edit`}
-                        className="rounded-md bg-accent px-2 py-1 text-xs"
-                      >
-                        Edit
-                      </a>
+                      {user?.role === "User" ? (
+                        <span
+                          className="rounded-md bg-accent px-2 py-1 text-xs text-muted-foreground opacity-75 cursor-not-allowed"
+                          title="You cannot edit tasks"
+                        >
+                          Edit
+                        </span>
+                      ) : (
+                        <a
+                          href={`/tasks/${task.id}/edit`}
+                          className="rounded-md bg-accent px-2 py-1 text-xs hover:opacity-90"
+                        >
+                          Edit
+                        </a>
+                      )}
+
                       <button
+                        disabled={user?.role === "User"}
                         onClick={() => handleDelete(task.id)}
-                        className="rounded-md bg-destructive/10 px-2 cursor-pointer py-1 text-xs"
+                        className={`rounded-md bg-destructive/10 px-2 py-1 text-xs ${
+                          user?.role === "User"
+                            ? "text-muted-foreground opacity-75 cursor-not-allowed"
+                            : "hover:opacity-90"
+                        }`}
                       >
                         Delete
                       </button>
